chatbot_project/pyproject.toml
[project]
name = "chatbot-project"
version = "0.1.0"
description = ""
authors = [{name = "Oleksii_K", email = "sumyultras88@gmail.com"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "openai (>=1.66.3,<2.0.0)",
    "aiogram (>=3.18.0,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "numpy (>=2.2.3,<3.0.0)",
    "faker (>=37.0.0,<38.0.0)",
    "nltk (>=3.9.1,<4.0.0)",
    "rank-bm25 (>=0.2.2,<0.3.0)",
    "whitenoise (>=6.9.0,<7.0.0)",
    "pgvector[django] (>=0.4.0,<0.5.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    # "djangorestframework (>=3.15.2,<4.0.0)",   # розкоментуй, якщо реально використовуєш DRF
    # "django-ratelimit (>=4.1.0,<5.0.0)",       # розкоментуй, якщо додаси rate limiting
    "pillow (>=11.3.0,<12.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pgcli = "^4.3.0"

